requirements.txt
tqdm==4.42.1
ratelimiter==1.2.0.post0
tencentcloud_sdk_python==3.0.441
